# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Portlyn API
  version: 0.1.0
  description: |
    Initial OpenAPI description for core Portlyn control-plane APIs.
    This spec focuses on auth, nodes, certificates, and DNS providers.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      summary: Login with local credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/request-otp:
    post:
      summary: Request one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: OTP request accepted
  /auth/verify-otp:
    post:
      summary: Verify OTP and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, token]
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
      responses:
        "200":
          description: OTP verification success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /nodes/enroll:
    post:
      summary: Enroll a node with enrollment token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, name]
              properties:
                token:
                  type: string
                name:
                  type: string
                description:
                  type: string
                version:
                  type: string
      responses:
        "201":
          description: Node enrolled
  /nodes/{id}/heartbeat:
    post:
      summary: Send node heartbeat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeHeartbeat"
      responses:
        "200":
          description: Heartbeat accepted
  /certificates:
    get:
      summary: List certificates
      responses:
        "200":
          description: Certificate list
    post:
      summary: Create certificate resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateCreate"
      responses:
        "201":
          description: Created
  /certificates/{id}/import-pem:
    post:
      summary: Import existing PEM/KEY for certificate resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [certificate_pem, private_key_pem]
              properties:
                certificate_pem:
                  type: string
                private_key_pem:
                  type: string
                issuer_key:
                  type: string
      responses:
        "200":
          description: PEM import success
  /dns-providers:
    get:
      summary: List DNS providers
      responses:
        "200":
          description: DNS provider list
    post:
      summary: Create DNS provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DNSProviderCreate"
      responses:
        "201":
          description: Created
  /dns-providers/{id}/test:
    post:
      summary: Validate stored DNS provider config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation success
components:
  schemas:
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        requires_mfa:
          type: boolean
        mfa_token:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
    NodeHeartbeat:
      type: object
      properties:
        version:
          type: string
        status:
          type: string
        load:
          type: number
        bandwidth_in_kbps:
          type: number
        bandwidth_out_kbps:
          type: number
    CertificateCreate:
      type: object
      required: [domain_id, type, challenge_type, issuer]
      properties:
        domain_id:
          type: integer
        type:
          type: string
          enum: [single, wildcard, multi_san]
        challenge_type:
          type: string
          enum: [http-01, dns-01]
        issuer:
          type: string
          enum: [letsencrypt_prod, letsencrypt_staging]
        primary_domain:
          type: string
        sans:
          type: array
          items:
            type: string
        renewal_window_days:
          type: integer
        is_auto_renew:
          type: boolean
        dns_provider_id:
          type: integer
    DNSProviderCreate:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cloudflare, hetzner, route53, digitalocean]
        config:
          type: object
          additionalProperties:
            type: string
        is_active:
          type: boolean